rename tesseract3 -> tesserocr
