+++ ocrd/processor/recognize/tesserocr.py
@@ -1,3 +1,5 @@
+from __future__ import absolute_import
+
 from ocrd.model import OcrdPage
 from ocrd.processor.base import Processor
 from ocrd.utils import getLogger, mets_file_id
