+++ ocrd/processor/segment_line/tesserocr.py
@@ -1,3 +1,5 @@
+from __future__ import absolute_import
+
 from ocrd.model import OcrdPage
 from ocrd.processor.base import Processor
 from ocrd.utils import getLogger, mets_file_id
@@ -5,7 +7,7 @@ from ocrd.constants import MIMETYPE_PAGE
 
 import tesserocr
 
-log = getLogger('processor.segment_line.tesseract3')
+log = getLogger('processor.segment_line.tesserocr')
 
 class Tesseract3LineSegmenter(Processor):
