+++ test/processor/segment_line/test_tesserocr.py
@@ -1,7 +1,7 @@
 #  from pprint import pprint
 from ocrd.resolver import Resolver
-from ocrd.processor.segment_region.tesseract3 import Tesseract3RegionSegmenter
-from ocrd.processor.segment_line.tesseract3 import Tesseract3LineSegmenter
+from ocrd.processor.segment_region.tesserocr import Tesseract3RegionSegmenter
+from ocrd.processor.segment_line.tesserocr import Tesseract3LineSegmenter
 
 from test.assets import METS_HEROLD_SMALL
 from test.base import TestCase, main
