+++ test/processor/segment_region/test_tesserocr.py
@@ -3,7 +3,7 @@
 from test.base import TestCase, main
 from test.assets import METS_HEROLD_SMALL
 from ocrd.resolver import Resolver
-from ocrd.processor.segment_region.tesseract3 import Tesseract3RegionSegmenter
+from ocrd.processor.segment_region.tesserocr import Tesseract3RegionSegmenter
 
 class TestProcessorSegmentRegionTesseract3(TestCase):
